+++ CHANGES.md
@@ -22,6 +22,7 @@ Release Notes.
 * Support Zabbix protocol to receive agent metrics.
 * Update the Apdex metric combine calculator.
 * Enhance `MeterSystem` to allow creating metrics with same `metricName` / `function` / `scope`.
+* Storage plugin supports postgresql.
 
 #### UI
 * Update selector scroller to show in all pages.